Account registrar chain authority
This account is responsible for adding new accounts into the blockchain.
All account registration operations must be signed by this authority
in order to be valid.

As with all chain authorities, there are committee operations that
can update (replace it) on a committee member proposal.

This will also update the genesis state and the default genesis.json to
contain the initial registrar. Note that this must be filled with an
account from the genesis.json and changed through committee proposal.